refactor: point 모듈에서 가격을 모두 long 타입으로 변환

+++ point/src/main/java/store/_0982/point/application/dto/PaymentPointCreateInfo.java
@@ -10,7 +10,7 @@ public record PaymentPointCreateInfo(
         UUID paymentPointId,
         UUID memberId,
         UUID orderId,
-        int amount,
+        long amount,
         PaymentPointStatus status,
         OffsetDateTime createdAt,
         OffsetDateTime requestedAt
